fix RangeFilter

+++ database/migrations/2024_09_02_134821_create_collectibles_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->string('article_number')->nullable();
             $table->string('condition');
             $table->foreignIdFor(Shelf::class)->constrained();
-            $table->unsignedInteger('purchase_price')->nullable();
+            $table->unsignedBigInteger('purchase_price')->nullable();
             $table->date('purchased_at')->nullable();
             $table->string('seller')->nullable();
             $table->string('additional_field')->nullable();
